# artifacts/api-server/src/routes/auth.ts
import { Router } from "express";
import bcrypt from "bcryptjs";
import { db } from "@workspace/db";
import { usersTable, brandsTable, reviewerProfilesTable } from "@workspace/db/schema";
import { eq } from "drizzle-orm";
import { signToken, requireAuth } from "../middlewares/auth";
import { validateBody, schemas } from "../middlewares/validate";
import { logEvent, EVENT_TYPES } from "../lib/events";

const router = Router();

router.post("/auth/register", validateBody(schemas.register), async (req, res) => {
  try {
    const { email: rawEmail, password, username, role, companyName } = req.body;
    const email = rawEmail.toLowerCase().trim();

    const existing = await db.select({ id: usersTable.id }).from(usersTable).where(eq(usersTable.email, email)).limit(1);
    if (existing.length > 0) {
      res.status(409).json({ error: "conflict", message: "Email already registered" });
      return;
    }

    const existingUsername = await db.select({ id: usersTable.id }).from(usersTable).where(eq(usersTable.username, username)).limit(1);
    if (existingUsername.length > 0) {
      res.status(409).json({ error: "conflict", message: "Username already taken" });
      return;
    }

    const passwordHash = await bcrypt.hash(password, 12);

    const user = await db.transaction(async (tx) => {
      const [newUser] = await tx.insert(usersTable).values({ email, passwordHash, username, role }).returning();

      if (role === "brand") {
        await tx.insert(brandsTable).values({ userId: newUser.id, companyName: companyName?.trim() || username });
      }

      await logEvent({ eventType: EVENT_TYPES.USER_REGISTER, actorId: newUser.id, entityType: "user", entityId: newUser.id, metadata: { role, email } }, tx);

      return newUser;
    });

    const token = signToken({ userId: user.id, email: user.email, username: user.username, role: user.role });

    res.status(201).json({
      token,
      user: { id: user.id, email: user.email, username: user.username, role: user.role, createdAt: user.createdAt },
    });
  } catch (err) {
    console.error(err);
    res.status(500).json({ error: "internal_error", message: "Registration failed" });
  }
});

router.post("/auth/login", validateBody(schemas.login), async (req, res) => {
  try {
    const { email: rawEmail, password } = req.body;
    const email = rawEmail.toLowerCase().trim();
    const [user] = await db.select().from(usersTable).where(eq(usersTable.email, email)).limit(1);

    if (!user) { res.status(401).json({ error: "unauthorized", message: "Invalid credentials" }); return; }

    const valid = await bcrypt.compare(password, user.passwordHash);
    if (!valid) { res.status(401).json({ error: "unauthorized", message: "Invalid credentials" }); return; }

    await logEvent({ eventType: EVENT_TYPES.USER_LOGIN, actorId: user.id, entityType: "user", entityId: user.id, metadata: { role: user.role } });

    const token = signToken({ userId: user.id, email: user.email, username: user.username, role: user.role });

    res.json({ token, user: { id: user.id, email: user.email, username: user.username, role: user.role, createdAt: user.createdAt } });
  } catch (err) {
    console.error(err);
    res.status(500).json({ error: "internal_error", message: "Login failed" });
  }
});

router.get("/auth/me", requireAuth, async (req, res) => {
  try {
    const [user] = await db.select().from(usersTable).where(eq(usersTable.id, req.user!.userId)).limit(1);
    if (!user) { res.status(404).json({ error: "not_found", message: "User not found" }); return; }

    await logEvent({ eventType: EVENT_TYPES.PROFILE_VIEWED, actorId: user.id, entityType: "user", entityId: user.id, metadata: { role: user.role } });

    let profile = null;
    if (user.role === "reviewer") {
      const [p] = await db.select().from(reviewerProfilesTable).where(eq(reviewerProfilesTable.userId, user.id)).limit(1);
      if (p) {
        profile = { gender: p.gender, ageBand: p.ageBand, state: p.state, employmentStatus: p.employmentStatus, educationLevel: p.educationLevel };
      }
    }

    res.json({ id: user.id, email: user.email, username: user.username, role: user.role, createdAt: user.createdAt, profile });
  } catch (err) {
    console.error(err);
    res.status(500).json({ error: "internal_error", message: "Failed to fetch profile" });
  }
});

// ─── GET /auth/profile ────────────────────────────────────────────────────────
router.get("/auth/profile", requireAuth, async (req, res) => {
  try {
    const [p] = await db.select().from(reviewerProfilesTable).where(eq(reviewerProfilesTable.userId, req.user!.userId)).limit(1);
    res.json(p ?? { userId: req.user!.userId, gender: null, ageBand: null, state: null, employmentStatus: null, educationLevel: null });
  } catch (err) {
    console.error(err);
    res.status(500).json({ error: "internal_error", message: "Failed to fetch profile" });
  }
});

// ─── PATCH /auth/profile ─────────────────────────────────────────────────────
router.patch("/auth/profile", requireAuth, async (req, res) => {
  try {
    const { gender, ageBand, state, employmentStatus, educationLevel } = req.body;
    const userId = req.user!.userId;

    const updateData: Record<string, unknown> = {};
    if (gender !== undefined) updateData.gender = gender;
    if (ageBand !== undefined) updateData.ageBand = ageBand;
    if (state !== undefined) updateData.state = state;
    if (employmentStatus !== undefined) updateData.employmentStatus = employmentStatus;
    if (educationLevel !== undefined) updateData.educationLevel = educationLevel;

    const [existing] = await db.select({ id: reviewerProfilesTable.id }).from(reviewerProfilesTable).where(eq(reviewerProfilesTable.userId, userId)).limit(1);

    let result;
    if (existing) {
      const [updated] = await db.update(reviewerProfilesTable)
        .set({ ...updateData, updatedAt: new Date() } as any)
        .where(eq(reviewerProfilesTable.userId, userId))
        .returning();
      result = updated;
    } else {
      const [inserted] = await db.insert(reviewerProfilesTable)
        .values({ userId, ...updateData } as any)
        .returning();
      result = inserted;
    }

    res.json({ gender: result.gender, ageBand: result.ageBand, state: result.state, employmentStatus: result.employmentStatus, educationLevel: result.educationLevel });
  } catch (err) {
    console.error(err);
    res.status(500).json({ error: "internal_error", message: "Failed to update profile" });
  }
});

export default router;