# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AdSpot API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email or username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reviews/start:
    post:
      operationId: startReview
      tags: [reviews]
      summary: Start a review session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartReviewRequest"
      responses:
        "201":
          description: Review session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewSession"
  /reviews/{sessionId}/complete:
    post:
      operationId: completeReview
      tags: [reviews]
      summary: Complete a review session with answers
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteReviewRequest"
      responses:
        "200":
          description: Review completed with points awarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteReviewResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all platform settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettingsResponse"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update platform settings (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettingsResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        username:
          type: string
          minLength: 3
        role:
          type: string
          enum: [reviewer, brand]
        companyName:
          type: string
      required:
        - email
        - password
        - username
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [reviewer, brand, admin, super_admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - username
        - role
        - createdAt
...
    ReviewSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        adId:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [in_progress, completed, abandoned]
      required:
        - id
        - userId
        - adId
        - startedAt
        - status
    StartReviewRequest:
      type: object
      properties:
        adId:
          type: string
          format: uuid
      required:
        - adId
    AnswerInput:
      type: object
      properties:
        questionId:
          type: string
          format: uuid
        answerText:
          type: string
          nullable: true
        answerValue:
          type: string
          nullable: true
      required:
        - questionId
    CompleteReviewRequest:
      type: object
      properties:
        watchSeconds:
          type: integer
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
      required:
        - watchSeconds
        - answers
    CompleteReviewResponse:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/ReviewSession"
        pointsAwarded:
          type: integer
        totalBalance:
          type: integer
      required:
        - session
        - pointsAwarded
        - totalBalance
...
    PlatformSetting:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        label:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [string, number, boolean]
        updatedAt:
          type: string
          format: date-time
      required:
        - key
        - value
        - label
        - type
        - updatedAt
    PlatformSettingsResponse:
      type: object
      properties:
        settings:
          type: array
          items:
            $ref: "#/components/schemas/PlatformSetting"
      required:
        - settings
    UpdateSettingsRequest:
      type: object
      properties:
        updates:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              value:
                type: string
            required:
              - key
              - value
      required:
        - updates
...